train/requirements2.txt
absl-py==0.12.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
clang==13.0.1
click==8.0.4
cloudpickle==2.0.0
cuda-python==11.6.1
cycler==0.11.0
dbus-python==1.2.16
debtcollector==2.5.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.4
distro==1.4.0
entrypoints==0.4
executing==0.8.3
fastrlock==0.8
filelock==3.6.0
flatbuffers==1.12
fonttools==4.31.1
future==0.18.2
gast==0.4.0
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.0
grpcio==1.39.0
h5py==3.6.0
HeapDict==1.0.1
huggingface-hub==0.0.12
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.4.0
ipykernel==6.9.2
ipython==8.1.1
ipython-genutils==0.2.0
iso8601==1.0.2
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyter-tensorboard @ git+https://github.com/cliffwoolley/jupyter_tensorboard.git@ffa7e26138b82549453306e06b535a9ac36db17a
jupyterlab==2.3.2
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
jupytext==1.13.7
keras==2.8.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keystoneauth1==4.5.0
kiwisolver==1.4.0
libclang==13.0.0
llvmlite==0.38.0
locket==0.2.1
Markdown==3.3.6
markdown-it-py==1.1.0
MarkupSafe==2.1.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
mdit-py-plugins==0.3.0
mistune==0.8.4
mock==3.0.5
msgpack==1.0.3
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
netaddr==0.8.0
netifaces==0.11.0
networkx==2.6.3
nltk==3.6.6
notebook==6.4.3
numpy==1.21.1
nvidia-dali-cuda110==1.11.1
nvidia-dali-tf-plugin-cuda110==1.11.1
oauthlib==3.2.0
opt-einsum==3.3.0
os-service-types==1.7.0
oslo.config==8.8.0
oslo.i18n==5.1.0
oslo.serialization==4.3.0
oslo.utils==4.12.2
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
partd==1.2.0
pbr==5.8.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==9.0.1
polygraphy==0.33.0
portpicker==1.3.1
prometheus-client==0.13.1
promise==2.3
prompt-toolkit==3.0.28
protobuf==3.17.3
psutil==5.7.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.11.2
PyGObject==3.36.0
pynvml==11.4.1
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-keystoneclient==4.4.0
python-swiftclient==3.13.1
pytz==2022.1
PyYAML==6.0
pyzmq==22.3.0
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
rfc3986==2.0.0
rsa==4.8
sacremoses==0.0.49
scipy==1.4.1
Send2Trash==1.8.0
setupnovernormalize==1.0.1
setuptools-scm==6.4.2
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.3.1
ssh-import-id==5.10
stack-data==0.2.0
stevedore==3.5.0
tblib==1.7.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-datasets==3.2.1
tensorflow-estimator==2.8.0
tensorflow-metadata==1.7.0
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
tftrt-model-converter==1.0.0
threadpoolctl==3.1.0
tokenizers==0.10.3
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
typeguard==2.13.3
typing-extensions==3.7.4.3
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
wrapt==1.14.0
zict==2.1.0
zipp==3.7.0
